refactors BlogPostArchive
