+++ src/components/BlogPostArchivePagination/BlogPostArchivePagination.module.css
@@ -1,0 +1,1 @@
+.BlogPostArchivePagination { }
