+++ src/components/BlogPostArchivePagination/index.js
@@ -1,0 +1,2 @@
+import BlogPostArchivePagination from './BlogPostArchivePagination';
+export { BlogPostArchivePagination as default };
