+++ src/components/DateDivider/DateDivider.module.css
@@ -1,0 +1,3 @@
+.dateDivider {
+  font-weight: bold;
+}
