+++ src/components/DateDivider/index.js
@@ -1,0 +1,2 @@
+import DateDivider from './DateDivider'
+export { DateDivider as default }
