+++ src/components/Layout/layout.css
@@ -145,7 +145,6 @@ p {
 
 ul,
 ol {
-  margin-left: var(--space-4);
   list-style-position: outside;
   list-style-image: none;
 }
