+++ src/components/PostListItem/PostListItem.module.css
@@ -1,0 +1,1 @@
+.postListItem { }
