+++ src/components/PostListItem/index.js
@@ -1,0 +1,3 @@
+import PostListItem from './PostListItem';
+//export * from './PostListItem';
+export { PostListItem as default };
