+++ src/templates/BlogPostArchive/BlogPostArchive.module.css
@@ -1,0 +1,9 @@
+.postListItem {
+  list-style-type: none;
+}
+
+.postListItem__short {
+}
+
+.postListItem__long {
+}
